+++ src/answers/index.ts
@@ -1,27 +1,42 @@
 import { z } from 'zod';
-import { DateAnswerSchema, DateRangeAnswerSchema } from './dateAnswers';
-import { FilteredSearchAnswerSchema, TypeaheadSearchAnswerSchema } from './graphQLAnswers';
-import { CheckboxesAnswerSchema, RadioButtonsAnswerSchema, SelectBoxAnswerSchema } from './optionBasedAnswers';
 import {
-  BooleanAnswerSchema,
-  CurrencyAnswerSchema,
-  EmailAnswerSchema,
-  NumberAnswerSchema,
-  NumberRangeAnswerSchema,
-  TextAnswerSchema,
-  TextAreaAnswerSchema,
-  URLAnswerSchema
-} from './primitiveAnswers';
-import { TableAnswerSchema } from './tableAnswers';
+  DateAnswerSchema, DateAnswerType,
+  DateRangeAnswerSchema, DateRangeAnswerType
+} from './dateAnswers';
+import {
+  FilteredSearchAnswerSchema, FilteredSearchAnswerType,
+  TypeaheadSearchAnswerSchema, TypeaheadSearchAnswerType
+} from './graphQLAnswers';
+import {
+  CheckboxesAnswerSchema, CheckboxesAnswerType,
+  RadioButtonsAnswerSchema, RadioButtonsAnswerType,
+  SelectBoxAnswerSchema, SelectBoxAnswerType
+} from './optionBasedAnswers';
+import {
+  BooleanAnswerSchema, BooleanAnswerType,
+  CurrencyAnswerSchema, CurrencyAnswerType,
+  NumberAnswerSchema, NumberAnswerType,
+  NumberRangeAnswerSchema, NumberRangeAnswerType
+} from './numberAnswers';
+import {
+  TableAnswerSchema, TableAnswerType
+} from './tableAnswers';
+import {
+  EmailAnswerSchema, EmailAnswerType,
+  TextAnswerSchema, TextAnswerType,
+  TextAreaAnswerSchema, TextAreaAnswerType,
+  URLAnswerSchema, URLAnswerType
+} from './textAnswers';
 import { QuestionTypesEnum } from '../questions';
 
 // reexport everything
 export * from './answer';
 export * from './dateAnswers';
 export * from './graphQLAnswers';
+export * from './numberAnswers';
 export * from './optionBasedAnswers';
-export * from './primitiveAnswers';
 export * from './tableAnswers';
+export * from './textAnswers';
 
 // Union of all possible answers
 export const AnyAnswerSchema = z.discriminatedUnion('type', [
@@ -64,3 +79,22 @@ export const AnswerSchemaMap: Record<z.infer<typeof QuestionTypesEnum>, z.ZodTyp
 
 // This will ensure that object validations are against the Zod schemas defined above
 export type AnyAnswerType = z.infer<typeof AnyAnswerSchema>;
+
+export interface AnswerTypeMap {
+  boolean: BooleanAnswerType,
+  checkBoxes: CheckboxesAnswerType,
+  currency: CurrencyAnswerType,
+  date: DateAnswerType,
+  dateRange: DateRangeAnswerType,
+  email: EmailAnswerType,
+  filteredSearch: FilteredSearchAnswerType,
+  number: NumberAnswerType,
+  numberRange: NumberRangeAnswerType,
+  radioButtons: RadioButtonsAnswerType,
+  selectBox: SelectBoxAnswerType,
+  table: TableAnswerType,
+  text: TextAnswerType,
+  textArea: TextAreaAnswerType,
+  typeaheadSearch: TypeaheadSearchAnswerType,
+  url: URLAnswerType
+}