+++ src/answers/tableAnswers.ts
@@ -3,15 +3,8 @@ import { AnswerSchema } from './answer';
 import { CheckboxesAnswerSchema, RadioButtonsAnswerSchema, SelectBoxAnswerSchema } from './optionBasedAnswers';
 import { DateAnswerSchema, DateRangeAnswerSchema } from './dateAnswers';
 import { FilteredSearchAnswerSchema, TypeaheadSearchAnswerSchema } from './graphQLAnswers';
-import {
-  BooleanAnswerSchema,
-  CurrencyAnswerSchema,
-  EmailAnswerSchema,
-  NumberAnswerSchema,
-  TextAreaAnswerSchema,
-  TextAnswerSchema,
-  URLAnswerSchema
-} from './primitiveAnswers';
+import { BooleanAnswerSchema, CurrencyAnswerSchema, NumberAnswerSchema } from './numberAnswers';
+import { EmailAnswerSchema, TextAnswerSchema, TextAreaAnswerSchema, URLAnswerSchema } from './textAnswers';
 
 // Answers to Table Column Question Types (note that TableAnswer is not included here because we don't allow nested tables)
 export const AnyTableColumnAnswerSchema = z.discriminatedUnion('type', [