+++ src/questions/question.ts
@@ -22,13 +22,22 @@ export const QuestionTypesEnum = z.enum([
   'url'
 ]);
 
+const DefaultMetaSchema = z.object({
+  schemaVersion: z.literal(CURRENT_SCHEMA_VERSION),       // The schema version (default is CURRENT_SCHEMA_VERSION)
+});
+
+const DefaultAttributesSchema = z.object({
+  label: z.string().optional(),                           // UI label for the field
+  help: z.string().optional(),                            // Help/tooltip text for the field
+
+  labelTranslationKey: z.string().optional()              // The translation key for the label (DMP Tool only)
+});
+
 // Base abstract type for all questions
 export const QuestionSchema = z.object({
-  type: QuestionTypesEnum,                                  // The type of question
-  meta: z.object({                                          // The metadata for the question
-    schemaVersion: z.literal(CURRENT_SCHEMA_VERSION),       // The schema version (default is CURRENT_SCHEMA_VERSION)
-    labelTranslationKey: z.string().optional()              // The translation key for the label (DMP Tool only)
-  }),
+  type: QuestionTypesEnum,                                // The type of question
+  attributes: DefaultAttributesSchema.optional(),         // Field attributes
+  meta: DefaultMetaSchema,                                // Meta information for the field
 });
 
 // This will ensure that object validations are against the Zod schemas defined above