+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # dmptool-types CHANGELOG
 
+## v1.1.0
+
+- Added `QuestionTypeMap` and `AnswerTypeMap` that maps the `type` label to the corresponding Type
+- Refactored `primitiveQuestions` into separate `numberQuestions` and `textQuestions`
+- Added `label` and `help` to the base attributes for every question type
+- Moved `denomination` from `meta` to `attributes` in the currency question type
+- Fixed some issues with the Zod structure that was causing fields like `min` and `max` on the number question type to be lost 
+
 ## v1.0.8
 
 - Added `NumberRangeQuestionType` and `NumberRangeAnswerType`